pyproject.toml
[project]
name = "cdm-data-loaders"
version = "0.1.0"
description = "Data loaders and wranglers for the CDM."
requires-python = ">= 3.13"
readme = "README.md"
authors = [
    {name = "KBase developers", email = "engage@kbase.us"},
]

dependencies = [
    "bioregistry>=0.13.20",
    "boto3[crt]>=1.42.55",
    "click>=8.3.1",
    "defusedxml>=0.7.1",
    "delta-spark>=4.1.0",
    "dlt[deltalake,filesystem,parquet]>=1.22.2",
    "lxml>=6.0.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "thefuzz>=0.22.1",
]

[project.scripts]
idmapping = "cdm_data_loaders.parsers.uniprot.idmapping:cli"
uniprot_pipeline = "cdm_data_loaders.pipelines.uniprot_kb_pipeline:cli"
uniref_pipeline = "cdm_data_loaders.pipelines.uniref_pipeline:cli"

[dependency-groups]
dev = [
    "berdl-notebook-utils",
    "biopython>=1.86",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-env>=1.5.0",
    "ruff>=0.15.0",
]
models = [
    "genson>=1.3.0",
    "json2python-models>=0.3.1",
]
xml = [
    "xmlschema>=4.3.1",
    "xsdata[cli,lxml]>=26.1",
]

[project.optional-dependencies]
# for minio interactions -- see utils/minio.py for more details
minio = [
    "boto3[crt]>=1.42.0",
    "tqdm>=4.67.3",
]

biopython = [
    "biopython>=1.86",
]

[tool.ruff]
line-length = 120
target-version = "py313"
# Exclude a variety of commonly ignored directories.
exclude = [
    "__pypackages__",
    "_build",
    ".bzr",
    ".direnv",
    ".eggs",
    ".git-rewrite",
    ".git",
    ".github",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "*.pyc",
    "buck-out",
    "build",
    "deps",
    "dist",
    "node_modules",
    "python-coverage",
    "venv",
]

[tool.ruff.lint]
select = [
    # core
    "F", # Pyflakes
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "C90", # mccabe +
    "I", # isort
    "N", # pep8-naming
    "D", # pydocstyle
    "UP", # pyupgrade
    # extensions
    "YTT", # flake8-2020
    "ANN", # flake8-annotations
    "ASYNC", # flake8-async
    "S", # flake8-bandit
    "BLE", # flake8-blind-except
    "FBT", # flake8-boolean-trap
    "B", # flake8-bugbear
    "A", # flake8-builtins
    # "COM", # flake8-commas
    # "CPY", # flake8-copyright
    "C4", # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T10", # flake8-debugger
    "DJ", # flake8-django
    "EM", # flake8-errmsg
    "EXE", # flake8-executable
    "FA", # flake8-future-annotations
    "ISC", # flake8-implicit-str-concat
    "ICN", # flake8-import-conventions
    "G", # flake8-logging-format
    "INP", # flake8-no-pep420
    "PIE", # flake8-pie
    "T20", # flake8-print
    "PYI", # flake8-pyi
    "PT", # flake8-pytest-style
    "Q", # flake8-quotes
    "RSE", # flake8-raise
    "RET", # flake8-return
    "SLF", # flake8-self
    "SLOT", # flake8-slots
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
    "TC", # flake8-type-checking
    "INT", # flake8-gettext
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "TD", # flake8-todos
    "FIX", # flake8-fixme
    "ERA", # eradicate
    "PD", # pandas-vet
    "PGH", # pygrep-hooks
    "PL", # Pylint
    "TRY", # tryceratops
    "FLY", # flynt
    "NPY", # NumPy-specific rules
    "AIR", # Airflow
    "PERF", # Perflint
    "FURB", # refurb
    "LOG", # flake8-logging
    "RUF", # Ruff-specific rules
]

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]

ignore = [
    # D200: unnecessary-multiline-docstring
    "D200",
    # D203: one-blank-line-before-class (conflicts with D211)
    "D203",
    # D212: multi-line-summary-first-line (conflicts with D213)
    "D212",
    # E203: whitespace before ',', ';', or ':'
    "E203",
    # E501: line length (specified elsewhere)
    "E501",
    # ISC001: conflicts with Ruff's formatter
    "ISC001",
]

[tool.ruff.lint.per-file-ignores]
"*.ipynb" = ["T201"] # ignore printing in notebooks
"tests/**/*.py" = ["S101", "T201", "FBT001", "FBT002"]  # use of assert
"**/__init__.py" = ["D104"]

[tool.ruff.lint.mccabe]
# Flag errors (`C901`) whenever the complexity level exceeds 15.
max-complexity = 15

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = ["uv_build>=0.9.9,<0.11.0"]
build-backend = "uv_build"

[tool.pytest]
pythonpath = ["src"]
log_cli = true
log_cli_level = "INFO"
log_level = "INFO"
addopts = ["-v"]
markers = ["requires_spark: must be run in an environment where spark is available", "slow_test: does what it says on the tin"]

# environment settings for running tests
[tool.pytest_env]
USER = "fake_user"
KBASE_AUTH_TOKEN = "test-token-123"
CDM_TASK_SERVICE_URL = "http://localhost:8080"
MINIO_ENDPOINT_URL = "http://localhost:9000"
MINIO_ACCESS_KEY = "minioadmin"
MINIO_SECRET_KEY = "minioadmin"
MINIO_SECURE_FLAG = "false"
BERDL_POD_IP = "192.168.1.100"
SPARK_MASTER_URL = "spark://localhost:7077"
BERDL_HIVE_METASTORE_URI = "thrift://localhost:9083"
SPARK_CLUSTER_MANAGER_API_URL = "http://localhost:8000"
GOVERNANCE_API_URL = "http://localhost:8000"
DATALAKE_MCP_SERVER_URL = "http://localhost:8080"

[tool.uv.sources]
cdm-schema = { git = "https://github.com/kbase/cdm-schema.git" }
berdl-notebook-utils = { git = "https://github.com/BERDataLakehouse/spark_notebook.git", subdirectory = "notebook_utils" }
